Refactor e2e cleanup behavior for SKIP_RESOURCE_CLEANUP and SkipDeletionTest (#1161)
* Refactor e2e cleanup


* Revert focus on test


* Add data to e2e/readme, refactor the behavior, move BootstrapArtifactDump to before charts uninstallation and management cluster deletion


---------

+++ test/e2e/suites/chart-upgrade/suite_test.go
@@ -22,10 +22,12 @@ package chart_upgrade
 import (
 	"context"
 	"encoding/json"
+	"strconv"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+
 	"github.com/rancher/turtles/test/e2e"
 	"github.com/rancher/turtles/test/testenv"
 
@@ -100,6 +102,17 @@ var _ = SynchronizedAfterSuite(
 	func() {
 	},
 	func() {
+		By("Dumping artifacts from the bootstrap cluster")
+		testenv.DumpBootstrapCluster(ctx)
+
+		config := e2e.LoadE2EConfig()
+		// skipping error check since it is already done at the beginning of the test in e2e.ValidateE2EConfig()
+		skipCleanup, _ := strconv.ParseBool(config.GetVariable(e2e.SkipResourceCleanupVar))
+		if skipCleanup {
+			// add a log line about skipping charts uninstallation and cluster cleanup
+			return
+		}
+
 		testenv.UninstallRancherTurtles(ctx, testenv.UninstallRancherTurtlesInput{
 			BootstrapClusterProxy: setupClusterResult.BootstrapClusterProxy,
 		})
